Reset inputs function & useForm hook added
