+++ src/component/form.tsx
@@ -1,11 +1,9 @@
-import { useState } from "react";
-import Swal from "sweetalert2";
-
 // import fbIlus from "./../assets/fb.svg";
 // import linkedinIlus from "./../assets/linkedin.svg";
 // import igIlus from "./../assets/instagram.svg";
 import ilusContact from "./../assets/ilus-contact.svg";
 import { ButtonComponent } from "./shared";
+
 import "../sass/component/_form.scss";
 
 import {
@@ -22,99 +20,28 @@ import {
   types_of_business,
   services_of_interest,
 } from "../utils";
-import { FormState, SelectOptions } from "../interfaces";
+import { useForm } from "../hooks";
 
 interface Props {
   showOverlay: boolean;
 }
 
 export default function FormContact({ showOverlay }: Props) {
-  const [formState, setFormState] = useState<FormState>({
-    first_name: "",
-    last_name: "",
-    email: "",
-    phone: "",
-    business_name: "",
-  });
-
-  const [options, setOptions] = useState<SelectOptions>({
-    business_type: "",
-    market_size: "",
-  });
-
-  const [services_requested, setServicesRequested] = useState<string[]>([]);
-
-  const { first_name, last_name, email, phone, business_name } = formState;
-  const { business_type, market_size } = options;
-
-  const handleSubmit = async (
-    e: React.FormEvent<HTMLFormElement | HTMLSelectElement>
-  ) => {
-    e.preventDefault();
-
-    if (showOverlay) return;
-
-    try {
-      const invalidService = services_requested.filter(
-        (service) => service === "asesoría legal"
-      );
-
-      if (invalidService.length !== 0) return;
-
-      if (
-        Object.values({ ...formState, ...options }).every(
-          (item) => item.trim() !== ""
-        ) === false
-      )
-        return;
-
-      if (services_requested.length === 0) return;
-
-      const body = { ...formState, ...options, services_requested };
-
-      const request = await fetch(
-        "https://backoffice-git-main-joelibaceta.vercel.app/api/contact.py",
-        {
-          method: "POST",
-          body: JSON.stringify(body),
-        }
-      );
-
-      const response: string = await request.text();
-
-      const responseValidation: string[] | null = response.match(/ok/i);
-
-      if (responseValidation && responseValidation[0] === "OK") {
-        Swal.fire({
-          icon: "success",
-          title: "¡Genial!",
-          text: "Estamos felices de que formes parte de Groway Backoffice, pronto un ejecutivo de ventas lo contactará para poder gestionar su solicitud.",
-        });
-      }
-    } catch (error: any) {
-      throw new Error(error);
-    }
-  };
-
-  const handleInputChange = ({
-    target,
-  }: React.ChangeEvent<HTMLInputElement>) => {
-    setFormState((prevState) => ({
-      ...prevState,
-      [target.name]: target.value,
-    }));
-  };
-
-  const handleServices = ({ target }: React.ChangeEvent<HTMLInputElement>) => {
-    if (target.checked) {
-      setServicesRequested((state) => [...state, target.name]);
-    } else {
-      const servicesFiltered = services_requested.filter(
-        (service) => service !== target.name
-      );
-      setServicesRequested(servicesFiltered);
-    }
-  };
+  const {
+    business_name,
+    business_type,
+    checkedState,
+    email,
+    first_name,
+    last_name,
+    market_size,
+    phone,
+    handleInputChange,
+    handleOptionsChange,
+    handleServices,
+    handleSubmit,
+    handleOnChange,
+  } = useForm(showOverlay);
 
   return (
     <div className="formComponent" id="form">
@@ -203,12 +130,8 @@ export default function FormContact({ showOverlay }: Props) {
                 <Label>Tipo de Negocio</Label>
                 <Select
                   value={business_type}
-                  onChange={({ target }) =>
-                    setOptions((prevState) => ({
-                      ...prevState,
-                      business_type: target.value,
-                    }))
-                  }
+                  name="business_type"
+                  onChange={handleOptionsChange}
                 >
                   {types_of_business.map(({ text, value }) => (
                     <Option key={value} value={value}>
@@ -230,12 +153,8 @@ export default function FormContact({ showOverlay }: Props) {
                 <Label>Facturación Mensual</Label>
                 <Select
                   value={market_size}
-                  onChange={({ target }) =>
-                    setOptions((prevState) => ({
-                      ...prevState,
-                      market_size: target.value,
-                    }))
-                  }
+                  name="market_size"
+                  onChange={handleOptionsChange}
                 >
                   {monthly_income.map(({ text, value }) => (
                     <Option key={value} value={value}>
@@ -255,7 +174,11 @@ export default function FormContact({ showOverlay }: Props) {
                     <Checkbox
                       id={name}
                       name={name}
-                      onChange={handleServices}
+                      checked={checkedState[index]}
+                      onChange={(e) => {
+                        handleOnChange(index);
+                        handleServices(e);
+                      }}
                       disabled={index === 2}
                     />
                     <Label htmlFor={name}>{name}</Label>
